Add rust blaster util to netconsd repo
Summary: A rust version of "blaster" util is added here. The purpose of the utility binary is to manually stress netconsd by blasting it with messages, some functionality has been put in a library so that a self-test module can then blast itself to verify that netconsd works (to be added in a next diff).

Reviewed By: alastor-erinyes

Differential Revision: D40006457

fbshipit-source-id: 2aa25a0a5ab7257e2d87571da6a3d9944d484d7b