Add rust blaster util to netconsd repo (#10)
Summary:
Pull Request resolved: https://github.com/facebook/netconsd/pull/10

A rust version of "blaster" util is added here. The purpose of the utility binary is to manually stress netconsd by blasting it with messages, some functionality has been put in a library so that a self-test module can then blast itself to verify that netconsd works (to be added in a next diff).

Reviewed By: alastor-erinyes

Differential Revision: D40006457

fbshipit-source-id: 56c39a3613e46e4e955cc4c6edc193e6ef4d1a59