Add selftest module to netconsd repo (#11)
Summary:
Pull Request resolved: https://github.com/facebook/netconsd/pull/11

Added a selftest module that can be used to validate a netconsd build. The validation is performed by sending messages to loopback, netconsd would then pass those messages to the module that can verify that the messages were the ones that it actually sent.

The validation has a lot of room for improvement, as it is now, it sends a batch of messages and does a few basic checks on the received messages (like the total count, source ip and text content). If someone would need to modify netconsd itself, this module could be a good base to build more specific tests on.

Reviewed By: thepacketgeek

Differential Revision: D40105407

fbshipit-source-id: f53f57791b357a26392355d65f12a9cb46d643d7