Move filtering into get_target
Summary:
If we filter a line from a host we haven't seen before, we end up creating a zero byte file.
Move the filtering into get_target, which is where files get created, and make write_log dependant on that returning success, indicating that a msg wasn't dropped.

bump to c++17 for std::optional

Reviewed By: michel-slm

Differential Revision: D86524058

fbshipit-source-id: 4186d4665e6759a819d90ed5e358ea47be976217